appapi/apps/requirements.txt
Flask==2.1.0
gunicorn>=20.0.0
requests==2.26.0
torch==1.9.1
diffusers==0.2.0
tqdm==4.62.3
tokenizers>=0.0.13
